+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 # Head
 
+- Fixed: minimum required loader-utils version is now ~1.0.0.
 - Removed: support `webpack` version `1`.
 
 # 2.0.0 - 2017-02-01